undo changes for INTERNAL_DAC
removed the change, as it seems that the original code can handle the full range of int16 -> uint16.

+++ src/AudioOutputI2S.cpp
@@ -298,8 +298,8 @@ bool AudioOutputI2S::ConsumeSample(int16_t sample[2])
     uint32_t s32;
     if (output_mode == INTERNAL_DAC)
     {
-      uint32_t l = ((int32_t)Amplify(ms[LEFTCHANNEL])) + 0x8000;
-      uint32_t r = ((int32_t)Amplify(ms[RIGHTCHANNEL])) + 0x8000;
+      int16_t l = Amplify(ms[LEFTCHANNEL]) + 0x8000;
+      int16_t r = Amplify(ms[RIGHTCHANNEL]) + 0x8000;
       s32 = ((r & 0xffff) << 16) | (l & 0xffff);
     }
     else
